Allow delete requests to have a body
I don't know if this bug has always existed, or maybe the GitHub
API has been extended since this gem was written. But there are
cases where a DELETE request needs to send a body, for example:

https://developer.github.com/v3/repos/contents/#delete-a-file

This adds a body param, but sets the default value to nil in
order to keep existing code from breaking.